feat(vault): extend VaultToken
