+++ packages/blue-sdk/src/token/index.ts
@@ -1,2 +1,5 @@
 export * from "./Token";
 export * from "./WrappedToken";
+export * from "./ConstantWrappedToken";
+export * from "./ExchangeRateWrappedToken";
+export * from "./VaultToken";
